virustotal_rs::groups

Struct GroupsClient

pub struct GroupsClient { /* private fields */ }

Client for managing VirusTotal groups

Implementations

impl GroupsClient

pub fn new(client: Client) -> Self

pub async fn get_group(&self, id: &str) -> Result<GroupResponse>

Get a group object by ID

Retrieves information about a group.

Arguments

id - Group ID

Example

let group = groups_client.get_group("group123").await?;

pub async fn update_group( &self, id: &str, updates: &GroupUpdateRequest, ) -> Result<GroupResponse>

Update a group object

Updates group attributes. Attributes not present in the request remain unchanged.

Arguments

id - Group ID

updates - Partial group object with attributes to update

Example

use virustotal_rs::groups::{GroupUpdateRequest, GroupUpdate, GroupUpdateAttributes};
let updates = GroupUpdateRequest {
    data: GroupUpdate::new(GroupUpdateAttributes {
            name: Some("New Group Name".to_string()),
            description: Some("Updated description".to_string()),
            ..Default::default()
        })
};
let updated_group = groups_client.update_group("group123", &updates).await?;

pub async fn get_administrators(&self, id: &str) -> Result<AdminsResponse>

Get administrators for a group

Returns a list of User objects who are administrators of the group.

Arguments

id - Group ID

Example

let admins = groups_client.get_administrators("group123").await?;

pub async fn add_administrators( &self, id: &str, user_emails: Vec<&str>, ) -> Result<()>

Grant group admin permissions to users

Adds users as administrators. User emails must be used. No admins are removed by using this endpoint.

Arguments

id - Group ID

user_emails - List of user email addresses

Example

let emails = vec!["admin1@example.com", "admin2@example.com"];
groups_client.add_administrators("group123", emails).await?;

pub async fn is_administrator(&self, id: &str, user_id: &str) -> Result<bool>

Check if a user is a group admin

Arguments

id - Group ID

user_id - User ID to check

Example

let is_admin = groups_client.is_administrator("group123", "user456").await?;

pub async fn remove_administrator(&self, id: &str, user_id: &str) -> Result<()>

Revoke group admin permissions from a user

Arguments

id - Group ID

user_id - User ID to remove admin privileges from

Example

groups_client.remove_administrator("group123", "user456").await?;

pub async fn get_users(&self, id: &str) -> Result<UsersResponse>

Get group users

Returns a list of users who are members of the group.

Arguments

id - Group ID

Example

let users = groups_client.get_users("group123").await?;

pub async fn is_member(&self, id: &str, user_id: &str) -> Result<bool>

Check if a user is a group member

Arguments

id - Group ID

user_id - User ID to check

Example

let is_member = groups_client.is_member("group123", "user456").await?;

pub async fn remove_user(&self, id: &str, user_id: &str) -> Result<()>

Remove a user from a group

Arguments

id - Group ID

user_id - User ID to remove from the group

Example

groups_client.remove_user("group123", "user456").await?;

pub async fn add_users(&self, id: &str, user_emails: Vec<&str>) -> Result<()>

Add users to a group

Adds users as members. User emails must be used. No users are removed by using this endpoint.

Arguments

id - Group ID

user_emails - List of user email addresses

Example

let emails = vec!["user1@example.com", "user2@example.com"];
groups_client.add_users("group123", emails).await?;

pub async fn get_relationship<T: for<'de> Deserialize<'de>>( &self, id: &str, relationship: &str, limit: Option<u32>, cursor: Option<&str>, ) -> Result<Collection<T>>

Get objects related to a group

Retrieves related objects for a group.

Arguments

id - Group ID

relationship - Relationship name (e.g., “users”, “administrators”, “graphs”)

limit - Maximum number of related objects to retrieve

cursor - Continuation cursor for pagination

Example

let graphs: virustotal_rs::objects::Collection<serde_json::Value> = groups_client.get_relationship("group123", "graphs", Some(20), None).await?;

pub async fn get_relationship_descriptors<T: for<'de> Deserialize<'de>>( &self, id: &str, relationship: &str, limit: Option<u32>, cursor: Option<&str>, ) -> Result<Collection<T>>

Get object descriptors related to a group

Same as get_relationship but returns only descriptors and context attributes instead of full objects.

Arguments

id - Group ID

relationship - Relationship name

limit - Maximum number of related objects

cursor - Continuation cursor

pub fn iter_relationship<T: for<'de> Deserialize<'de> + Clone + 'static>( &self, id: &str, relationship: &str, ) -> CollectionIterator<'_, T>

Get an iterator for group relationships